🏗️ restructure brokerage
