+++ tests/conftest.py
@@ -1,9 +1,9 @@
 import pytest
 
+from pyrb.brokerage.base.fetcher import CurrentPrice, PriceFetcher
+from pyrb.brokerage.base.order_manager import Order, OrderManager
+from pyrb.brokerage.base.portfolio import Portfolio, Position
 from pyrb.brokerage.context import RebalanceContext
-from pyrb.brokerage.fetcher import CurrentPrice, PriceFetcher
-from pyrb.brokerage.order_manager import Order, OrderManager
-from pyrb.brokerage.portfolio import Portfolio, Position
 
 
 class FakePortfolio(Portfolio):
