+++ tests/test_main.py
@@ -1,8 +1,8 @@
 from pytest_mock import MockerFixture
 from typer.testing import CliRunner
 
+from pyrb.brokerage.base.order_manager import Order, OrderType
 from pyrb.brokerage.context import RebalanceContext
-from pyrb.brokerage.order_manager import Order, OrderType
 from pyrb.controller import app
 from pyrb.exceptions import InsufficientFundsException
